docker-dotpy-version/requirements.txt
apache-sedona==1.2.1
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
exceptiongroup==1.0.4
Fiona==1.8.22
geopandas==0.12.1
greenlet==2.0.1
iniconfig==1.1.1
munch==2.5.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
py4j==0.10.9
pyparsing==3.0.9
pyproj==3.4.0
pyspark==3.0.3
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
